Update patch set 3
Patch Set 3:

The job triggered at upload does not run the sonar job. Issuing the magic work to run the sonar job only run the sonar job using the merged code, not the code in the current (this) commit. So not possible to verify that the sonar job is ok before merge?
This commit works on a Ubuntu VM when running 'tox' from the command line.

Patch-set: 3