Create an external web server building RESTful API
-A Python server building REST supports HTTP/HTTPS protocols
-Reviewed for patchset2
-patch set 3 : Create a secondary docker image and move external server
inside in it
-patch set 4 : Create Open API specs and move base API operations inside
in it, change json schema of the test_message to a basic json file
-patch set 5 : Implement set|reset forced response and forced delay
including negative unit tests as well as get API for a single policy.
Code review cleanup
-patch set 6 : Change API URL of admin functions, add a test case to
check if an request lasts for an amount of time, and  clean-up.
-patch set 7 : Remove container-tag.yaml.
-patch set 8 : Make forced delay test case parametric function which
expected delay time in seconds can be passed in.
-patch set 9 : Fix typo.
-patch set 10: Generate my own certificate and key.

Issue-ID: NONRTRIC-755
Change-Id: I0ee622ad82e18639e4eb2d4a02863904f9bf3201
Signed-off-by: halil.cakal <halil.cakal@est.tech>